2sequentialAgent_health_insurer_acp/pyproject.toml
[project]
name = "2sequentialagent-health-insurer-acp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "acp-sdk>=1.0.1",
    "colorama>=0.4.6",
    "crewai>=0.150.0",
    "crewai-tools>=0.58.0",
    "duckduckgo-search>=8.1.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-openai>=0.2.14",
    "langgraph>=0.6.1",
    "litellm>=1.74.3",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.1.1",
]
